// frontend/src/api/model/playlist.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { PlaylistDescription } from './playlistDescription';

export interface Playlist {
  /** The name of the playlist. */
  name: string;
  /** Optional description of the playlist. */
  description?: PlaylistDescription;
  /** Unique database identifier for the playlist. */
  id: number;
  /** Timestamp when the playlist was created. */
  date_created: string;
  /** Number of images in the playlist. */
  image_count?: number;
}